requirements.txt
gradio
transformers
faiss-cpu
sentence-transformers
torch
openai-whisper
numpy
scipy
onnxruntime
soundfile
langdetect
gtts
